A board is now just data
