+++ bin/play.rb
@@ -1,7 +1,5 @@
 #!/usr/bin/env ruby
 require './lib/tic_tac_toe.rb'
-
-board = Board.new
-game = TicTacToe.new(board)
+game = TicTacToe.new
 game.play
